Default cluster for data which doesn't have category
